PM-439 Move actions, reducers related to provider integrations to src/integrations; Start working on using immutable implementation
