+++ .eslintrc
@@ -24,7 +24,8 @@
     "import/extensions": 0,
     "import/prefer-default-export": 0,
     "jsx-a11y/label-has-for": 0,
-    "indent": ["error", 2]
+    "indent": ["error", 2],
+    "no-console": ["error", { "allow": ["warn", "error"] }]
   },
   "env": {
     "jest/globals": true,
