+++ package.json
@@ -47,6 +47,7 @@
     "d3-scale-chromatic": "^1.1.1",
     "decimal.js": "^9.0.1",
     "history": "^4.7.2",
+    "immutable": "^3.8.2",
     "lodash": "^4.17.5",
     "moment": "^2.19.2",
     "moment-duration-format": "^2.2.1",
