+++ src/components/Header/MenuAccountDropdown/LogOut.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 
 import {
   logoutProvider,
-} from 'actions/providers'
+} from 'integrations/store/actions'
 
 const LogOut = ({ logout }) => (
   <li key="action-logout" className="menuAccountAction__action">
