+++ src/containers/DashboardPage/index.js
@@ -6,7 +6,7 @@ import { getMarkets, getAccountPredictiveAssets } from 'selectors/market'
 import { getAccountTrades } from 'selectors/marketTrades'
 import { getAccountShares } from 'selectors/marketShares'
 import { isGnosisInitialized } from 'selectors/blockchain'
-import { getCurrentAccount, getEtherTokensAmount, checkWalletConnection } from 'integrations/selectors'
+import { getCurrentAccount, getEtherTokensAmount, checkWalletConnection } from 'integrations/store/selectors'
 import { requestMarkets, requestAccountTrades, requestAccountShares, redeemWinnings } from 'actions/market'
 import { requestGasPrice, requestEtherTokens } from 'actions/blockchain'
 import { weiToEth } from 'utils/helpers'
