+++ src/containers/MarketDetailPage/index.js
@@ -19,7 +19,7 @@ import { getMarketTradesForAccount } from 'selectors/marketTrades'
 import { getMarketShares } from 'selectors/marketShares'
 import { getMarketGraph } from 'selectors/marketGraph'
 import { getGasCosts, getGasPrice, isGasCostFetched, isGasPriceFetched } from 'selectors/blockchain'
-import { checkWalletConnection, getCurrentAccount, getCurrentBalance, isOnWhitelist } from 'integrations/selectors'
+import { checkWalletConnection, getCurrentAccount, getCurrentBalance, isOnWhitelist } from 'integrations/store/selectors'
 import { isModerator, getModerators } from 'utils/helpers'
 
 let marketId
