+++ src/selectors/marketGraph.js
@@ -3,7 +3,7 @@ import { createSelector } from 'reselect'
 import { normalizeScalarPoint, getOutcomeName } from 'utils/helpers'
 import { OUTCOME_TYPES } from 'utils/constants'
 
-import { getMarketTrades } from './marketTrades'
+import { getMarketTrades } from 'selectors/marketTrades'
 
 const getFirstGraphPoint = (market) => {
   let firstPoint
